requirements.txt
Pillow==8.4.0
diffusers==0.14.0
torch==1.13.1
numpy
gradio==3.23.0
transformers
xformers
accelerate
imageio